change VarStore; refactor Expr/VarStore;
