+++ src/expr.rs
@@ -1,51 +1,64 @@
-use std::{fs::File, io::Write, path::Path};
+use std::fmt::{Debug, Formatter};
 
-pub struct VarStore(Vec<String>);
+use crate::types::Atom;
 
-impl VarStore {
-    pub fn new() -> Self {
-	VarStore(Vec::<String>::new())
-    }
-    /// Returns a list of variable names.
-    pub fn n_var(&self) -> usize {
-	self.0.len()
-    }
-    /// Returns `Some(var_name)` if `(var_name, index): (String, usize)` is contained in [`Self`]
-    pub fn try_get_by_index(&self, index: usize) -> Option<&String> {
-	self.0.get(index)
-    }
-    /// Returns `Some(index)` if `(var_name, index): (String, usize)` is contained in [`Self`]
-    pub fn try_get_by_string(&self, var_name: &String) -> Option<usize> {
-	self.0.iter().position(|x| x == var_name).and_then(|x| Some(x+1))
-    }
-    /// Checks if [`VarStore`] contains `var_name`
-    pub fn contains(& self, var_name: &String) -> bool{
-	self.0.contains(var_name)
-    }
-    /// Adds `var_name` to [`VarStore`]
-    pub fn insert(&mut self, var_name: String) {
-	if !self.contains(&var_name) {
-	    self.0.push(var_name);
+/// Representation of the binary operators.
+#[derive(Clone, Copy, Debug,Eq,PartialEq)]
+pub enum BiOp {
+    And,
+    Or,
+}
+
+/// Representation of a boolean expression with their sequence of operations.
+#[derive(Clone, PartialEq, Eq)]
+pub enum Expr {
+    Atom(Atom),
+    BiOp(Box<Self>, BiOp, Box<Self>),
+    Not(Box<Self>),
+}
+
+impl Debug for Expr {
+    fn fmt(&self, f: &mut Formatter<'_>) -> Result<(), std::fmt::Error> {
+	match self {
+	    Expr::Atom(atom) => f
+		.debug_tuple("Atom")
+		.field(atom)
+		.finish(),
+	    Expr::BiOp(x, op, y) => f
+		.debug_tuple("BiOp")
+		.field(x)
+		.field(op)
+		.field(y)
+		.finish(),
+	    Expr::Not(x) => f
+		.debug_tuple("Not")
+		.field(x)
+		.finish(),
 	}
-    }
-    /// Adds extra variable to [`VarStore`]. Returns index of variable.
-    pub fn new_extra_var(&mut self) -> i32 {
-	let new_var_index = self.n_var() + 1;
-	self.insert(format!("EXTRA_VAR_{}",new_var_index).to_string());
-	new_var_index as i32
-    }
-    /// Outputs `csv` file to [`VarStore`]
-    pub fn to_csv_file(&self, filename: &str) {
-	let path = Path::new(filename);
-        let mut file = File::create(path).expect("creating file failed");
-
-        let content = self.0.iter().enumerate().fold(
-	    String::new(),
-	    |mut old, (index,x)| {
-		old.push_str(&format!("{} {}\n", index + 1, x));
-		old
-	    });
-        file.write_all(content.as_bytes()).unwrap();
     }
 }
 
+impl From<Expr> for String {
+    fn from(value: Expr) -> Self {
+        match value {
+            Expr::BiOp(x, BiOp::And, y) => {
+                let x: String = (*x).into();
+                let y: String = (*y).into();
+                format!("({} & {})", x, y)
+            }
+            Expr::BiOp(x, BiOp::Or, y) => {
+                let x: String = (*x).into();
+                let y: String = (*y).into();
+                format!("({} | {})", x, y)
+            }
+            Expr::Not(x) => {
+                let x: String = (*x).into();
+                format!("!({})", x)
+            }
+	    // to be fixed
+            Expr::Atom(Atom::Var(x)) => format!("{}",x),
+            Expr::Atom(Atom::False) => "False".to_string(),
+            Expr::Atom(Atom::True) => "True".to_string(),
+        }
+    }
+}
