+++ src/lib.rs
@@ -1,7 +1,7 @@
 pub mod types;
 pub mod expr;
+pub mod var;
 pub mod lexer;
 pub mod parser;
 pub mod algorithm;
-pub mod transform;
 pub mod cnf;
